Fix size_t
https://github.com/kozyilmaz/zcash-apple/commit/c1299fd5fc1d2e7eff38fcd4f79ec26012eb81b5#diff-c687e3c1218feff0886a6ffb4674aed0

+++ src/rpcblockchain.cpp
@@ -684,7 +684,7 @@ UniValue getblockchaininfo(const UniValue& params, bool fHelp)
 
     ZCIncrementalMerkleTree tree;
     pcoinsTip->GetAnchorAt(pcoinsTip->GetBestAnchor(), tree);
-    obj.push_back(Pair("commitments",           tree.size()));
+    obj.push_back(Pair("commitments",           (uint64_t)tree.size()));
 
     const Consensus::Params& consensusParams = Params().GetConsensus();
     CBlockIndex* tip = chainActive.Tip();
